added one fully connected layer

+++ conv_net_sentence.py
@@ -303,8 +303,8 @@ def make_idx_data_cv(revs, word_idx_map, cv, max_l=51, k=300, filter_h=5):
             test.append(sent)
         else:
             train.append(sent)
-    train = np.array(train,dtype="int")
-    test = np.array(test,dtype="int")
+    train = np.array(train)
+    test = np.array(test)
     return [train, test]     
 
 def make_idx_data_tdt(revs, word_idx_map, max_l=51, k=300, filter_h=5):
